requirements.txt
discord
requests
requests-oauthlib
python-dotenv